change: Seed SpaceWeatherCachingProvider with local data if not initialized. Add download retry and backoff
